+++ src/main/java/dk/cngroup/lentils/controller/client/HintController.java
@@ -11,7 +11,6 @@ import dk.cngroup.lentils.service.CypherService;
 import dk.cngroup.lentils.service.GameLogicService;
 import dk.cngroup.lentils.service.HintService;
 import dk.cngroup.lentils.service.HintTakenService;
-import dk.cngroup.lentils.service.HintsTakenWithImageUrlService;
 import dk.cngroup.lentils.service.ScoreService;
 import dk.cngroup.lentils.service.StatusService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -41,7 +40,6 @@ public class HintController {
     private final CypherService cypherService;
     private final HintService hintService;
     private final HintTakenService hintTakenService;
-    private final HintsTakenWithImageUrlService hintsTakenWithImageUrlService;
     private final StatusService statusService;
     private final ScoreService scoreService;
     private final GameLogicService gameLogicService;
@@ -51,14 +49,12 @@ public class HintController {
                           final StatusService statusService,
                           final HintService hintService,
                           final HintTakenService hintTakenService,
-                          final HintsTakenWithImageUrlService hintsTakenWithImageUrlService,
                           final ScoreService scoreService,
                           final GameLogicService gameLogicService) {
         this.cypherService = cypherService;
         this.hintService = hintService;
         this.hintTakenService = hintTakenService;
         this.statusService = statusService;
-        this.hintsTakenWithImageUrlService = hintsTakenWithImageUrlService;
         this.scoreService = scoreService;
         this.gameLogicService = gameLogicService;
     }
@@ -106,7 +102,7 @@ public class HintController {
         model.addAttribute(TEMPLATE_ATTR_TEAM, user.getTeam());
         model.addAttribute(TEMPLATE_ATTR_CYPHER, cypher);
         model.addAttribute(TEMPLATE_ATTR_HINTS_TAKEN,
-                hintsTakenWithImageUrlService.addImageUrlsToHints(user.getTeam(), cypher));
+                hintTakenService.addImageUrlsToHints(user.getTeam(), cypher));
         model.addAttribute(TEMPLATE_ATTR_HINTS_NOT_TAKEN,
                 hintService.getAllNotTakenByTeamAndCypher(user.getTeam(), cypher));
         model.addAttribute(TEMPLATE_ATTR_SCORE, scoreService.getScoreByTeam(user.getTeam()));
